requirements.txt
discord.py
asyncio
table2ascii
bs4
pandas